Fix missing patch version in PHP 7.4 Unbuntu image tags.
Problem:
The PHP 7.4 Ubuntu images are being tagged with the patch missing. So what should be 7.4.2-laravel6 is ending up as 7.4.-laravel6.

Cause:
Prior to 7.4 the PHP version was being suffixed with something similar to "-6+ubuntu18.04.1+deb.sury.org+1". The patch extraction was stripping this out by dropping everything after the slash. When the suffix was no longer present, the stripping was returning an empty string.

Fix:
Conditionally check for the dash, and only strip if it is present.