Fix directives are not set if config files are symlinks
Problem:
When the entrypoint script tries to set PHP directives for the specified ENV vars, it only affects actual files, not symlinks. So pretty much just php.ini

Cause:
Ubuntu uses symlinks to link /etc/php/7.4/cli/conf.d/* to /etc/php/7.4/mods-available/*
Linux find does not follow symlinks by default

Fix:
Pass -follow to find so it follows the symlinks.
This has the added benefit of causing the symlinks to be replace with normal files when sed is executed on them. This is beneficial because it means that if you specify a CLI only directive, it will not affect FPM because the new file will be written to /etc/php/7.4/cli instead of modifying the symlink source file in /etc/php/7.4/mods-available and thus affecting FPM also.