Simplify tags
Problem:
Tags had a complicated construction, with alpha chars to distinguish between new images that use the same PHP version.

This resulted in tags that weren't ideal IMHO, as the alpha char would end up in the middle of the tag. e.g. with Laravel 10, if I created a new image that used PHP 8.3.3, and there was already an image for this version, the resulting tag would be chekote/php:8.3.3.a-laravel10.x, which made it seem as if the PHP version had some difference.

Also, this would only be done conditionally if the image had a different SHA. The SHA changes so often due to minor changes in file dates, etc, that it was pretty much always being created. I will address this particular concern in an upcoming commit.

Fix:
Instead of using the alpha char, always create a version of the image that has a datetime suffix. So using the PHP 8.3.3 + Laravel10 example again, we would have:

  php8-laravel10.x
  php8.1-laravel10.x
  php8.1.27-laravel10.x
  php8.1.27-laravel10.x-2024-03-08-13-59-20

Which means that even the patch version will now be overwritten, and if users want an immutable image tag they would need to use the one with the datetime suffix instead.