Push images to separate repos
Problem:
All of the images were being pushed to chekote/php.
This caused issues with automated tools such as Dependabot, as they would not be able to tell what newer image it should update to. It would simply choose the latest, which could result in a project using a Behat image being updated to a Laravel image (or visa versa)

Fix:
Re-organize the tags so that they are within repos that contain logical groupings. For example, Images for Laravel 10 will be exclusively in the chekote/laravel10.x repo.