Fix build stops at Laravel 11.x with PHP 8.4
Problem:
The build fails when attempting to build the Laravel 11.x image with PHP
8.4 with an error stating that the image is not compatible with PHP
above 8.3.*.

Cause:
The build script is building Laravel 11.x for all PHP versions. This was
not changed when PHP 8.4 support was added to the build script.

Fix:
Wrap the Laravel 11.x build in a conditional that skips it if we are
building with PHP higher than 8.3.*.